--- a/generator/beetle.css
+++ b/generator/beetle.css
@@ -1,24 +1,25 @@
-html, body {
-  margin: 0;
+html,
+body {
+    margin: 0;
 }
 
 html {
-  /*background: #fff url('wallpaper.jpg');*/
-  background-size: cover;
-  color: #333;
-  font: 16px/1.4 Open Sans, Helvetica, Arial, sans-serif;
+    /*background: #fff url('wallpaper.jpg');*/
+    background-size: cover;
+    color: #333;
+    font: 16px/1.4 Open Sans, Helvetica, Arial, sans-serif;
 }
 
 body {
-  max-width: 90%;
-  margin: 0 auto;
-  padding: 40px 0 0 0;
-  background-color: whiteSmoke;
-  display: flex;
-  justify-content: space-around;
+    max-width: 90%;
+    margin: 0 auto;
+    padding: 40px 0 0 0;
+    background-color: whiteSmoke;
+    display: flex;
+    justify-content: space-around;
 }
 
-button{
+button {
     background-color: white;
     color: black;
     border: none;
@@ -28,86 +29,86 @@ button{
     margin-bottom: 5px;
 }
 
-select {      
-      background-color: white;
-      color: black;
-      border: none;
-      padding: 3px;
-      box-shadow: 0 2px 5px rgba(0, 0, 0, .2);
-      margin-bottom: 10px;
+select {
+    background-color: white;
+    color: black;
+    border: none;
+    padding: 3px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, .2);
+    margin-bottom: 10px;
 }
 
-.cont > button{
+.cont>button {
     width: 100%;
     padding: 10px;
 }
 
-fieldset > button{
+fieldset>button {
     width: 100%;
 }
 
-#menue > button{
+#menue>button {
     padding: 8px;
 }
 
-#name{
+#name {
     width: 50%
 }
 
 input[type="color"] {
-	-webkit-appearance: none;
-	border: none;
+    -webkit-appearance: none;
+    border: none;
 }
+
 input[type="color"]::-webkit-color-swatch-wrapper {
-	padding: 0;
+    padding: 0;
 }
+
 input[type="color"]::-webkit-color-swatch {
-	border: none;
+    border: none;
 }
 
 input[type=range] {
-  width: 100%; /* Specific width is required for Firefox. */
-  background: transparent; /* Otherwise white in Chrome */
+    width: 100%;
+    /* Specific width is required for Firefox. */
+    background: transparent;
+    /* Otherwise white in Chrome */
 }
 
 input[type=range]:focus {
-  outline: none; /* Removes the blue border. You should probably do some kind of focus styling for accessibility reasons though. */
+    outline: none;
+    /* Removes the blue border. You should probably do some kind of focus styling for accessibility reasons though. */
 }
 
 input[type=range]::-ms-track {
-  width: 100%;
-  cursor: pointer;
-
-  /* Hides the slider so custom styles can be added */
-  background: transparent; 
-  border-color: transparent;
-  color: transparent;
+    width: 100%;
+    cursor: pointer;
+    /* Hides the slider so custom styles can be added */
+    background: transparent;
+    border-color: transparent;
+    color: transparent;
 }
 
 main {
     flex-basis: 800px;
-    
 }
 
-#smartomat{
+#smartomat {
     max-width: 900px;
-    
 }
 
-#sshop{
-    
-}
+#sshop {}
 
 #beetle {
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-size: cover;
-  min-height: 500px;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, .2);
-  background-position: center center;
-  background-repeat: no-repeat;
+    overflow: hidden;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-size: cover;
+    min-height: 500px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, .2);
+    background-position: center center;
+    background-repeat: no-repeat;
 }
 
 /*#beetle:not(.paused) {
@@ -115,121 +116,115 @@ main {
 }*/
 
 @keyframes breath {
-  0% { background-size: 100% 100%; }
-  100% { background-size: 120% 120%; }  
+    0% {
+        background-size: 100% 100%;
+    }
+    100% {
+        background-size: 120% 120%;
+    }
 }
 
 #beetlecanvas {
-  max-width: 70%;
-  min-height: 100px;
-  max-height: 400px;
-  cursor: grab;
+    max-width: 70%;
+    min-height: 100px;
+    max-height: 400px;
+    cursor: grab;
 }
 
 .meta {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  padding: 2em 2em 1.7em;
-  align-items: center;
-  background-color: white;
-  margin-top: 40px;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    padding: 2em 2em 1.7em;
+    align-items: center;
+    background-color: white;
+    margin-top: 40px;
 }
 
-main > button{
+main>button {
     width: 100%;
 }
 
-.design{
-
+.design {
     max-width: 300px;
 }
 
-.icon{
+.icon {
     width: 40px;
 }
 
-button:hover{
+button:hover {
     background-color: lightSteelBlue;
     cursor: pointer;
 }
 
 #name {
-  flex: 1 1 0;
-  text-align: left;
-  font-size: 1.4em;
-  color: #444;
-  font-family: serif;
-  line-height: 1;
+    flex: 1 1 0;
+    text-align: left;
+    font-size: 1.4em;
+    color: #444;
+    font-family: serif;
+    line-height: 1;
 }
 
 #refresh a {
-  color: #666;
-  text-decoration: none;
-  font-weight: 200;
-  margin-right: 30px;
+    color: #666;
+    text-decoration: none;
+    font-weight: 200;
+    margin-right: 30px;
 }
 
 #seed a {
-  color: #666;
-  text-decoration: none;
-  font-weight: 200;
+    color: #666;
+    text-decoration: none;
+    font-weight: 200;
 }
 
-
-
 #defaultCanvas0 {
-  display: none;
+    display: none;
 }
 
 footer {
-  text-align: center;
-  font-size: 0.9em;
-  margin: 0 0 1em;
-  color: #666;
-  text-shadow: 0 0 2px #fff;
-  position: absolute;
-  top: 20px;
-  left: 20px;
+    text-align: center;
+    font-size: 0.9em;
+    margin: 0 0 1em;
+    color: #666;
+    text-shadow: 0 0 2px #fff;
+    position: absolute;
+    top: 20px;
+    left: 20px;
 }
 
 footer a {
-  color: #444;
+    color: #444;
 }
 
-fieldset{
+fieldset {
     background-color: white;
-    display:none;
-    margin-bottom:15px;
+    display: none;
+    margin-bottom: 15px;
     border: none;
     box-shadow: 0 2px 5px rgba(0, 0, 0, .2);
 }
-    
-#wrap{
+
+#wrap {
     flex-basis: 400px;
     margin: 1em;
 }
 
-.cont{
+.cont {
     width: 100%;
-    
-}
-    
-
-
-#shirt-container{
-    
-    
-    
 }
 
-.container1{
-    width:1500px;
+#shirt-container {}
+
+.container1 {
+    width: 1500px;
     height: 700px;
 }
 
-.container2{
-    width:1500px;
+.container2 {
+    width: 1500px;
     height: 1500px;
     position: relative;
     left: 0px;
@@ -238,8 +233,8 @@ fieldset{
     justify-content: space-around;
 }
 
-.container0{
-    width:500px;
+.container0 {
+    width: 500px;
     height: 500px;
     position: relative;
     left: 0px;
@@ -248,7 +243,7 @@ fieldset{
     justify-content: space-around;
 }
 
-.singlebeetle{
+.singlebeetle {
     width: 300px;
     height: 300px;
     display: flex;
@@ -256,7 +251,7 @@ fieldset{
     background-size: contain;
 }
 
-.singlebeetle0{
+.singlebeetle0 {
     width: 500px;
     height: 500px;
     display: flex;
@@ -264,33 +259,30 @@ fieldset{
     background-size: contain;
 }
 
-.rasterimg{
+.rasterimg {
     margin: auto 0;
 }
 
-#shirt-background{
+#shirt-background {
     height: 400px;
-    background-position: center; 
+    background-position: center;
     position: relative;
     top: 150px;
     background-color: white;
 }
 
-#shirt-canvas{
-    
+#shirt-canvas {
     height: 700px;
     position: relative;
     top: -400px;
     margin: 0 auto;
-    
-
 }
 
-.shirt-body{
+.shirt-body {
     max-width: 100% !important;
 }
 
-#cont{
+#cont {
     display: flex;
     justify-content: space-evenly;
     flex-wrap: wrap;
@@ -299,29 +291,29 @@ fieldset{
 
 @keyframes spin {
     from {
-        transform:rotate(0deg);
+        transform: rotate(0deg);
     }
     to {
-        transform:rotate(360deg);
+        transform: rotate(360deg);
     }
 }
 
-#spinner{
+#spinner {
     animation-name: spin;
     animation-duration: 4000ms;
     animation-iteration-count: infinite;
     animation-timing-function: linear;
     width: 200px;
-    
 }
 
-.wait{
+.wait {
     margin: 0 auto;
     width: 300px;
 }
 
-.window{
-    max-width: 700px; !important
+.window {
+    max-width: 700px;
+     !important
 }
 
 /*
@@ -353,165 +345,165 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
 #beetle:not(.paused) #beetlecanvas:hover {
-  animation-name: shake-hard;
-  animation-duration: 3500ms;
-  animation-timing-function: ease-in-out;
-  animation-iteration-count: infinite;
-  animation-play-state: running;
+    animation-name: shake-hard;
+    animation-duration: 3500ms;
+    animation-timing-function: ease-in-out;
+    animation-iteration-count: infinite;
+    animation-play-state: running;
 }
 
 @keyframes shake-hard {
- 2% {
-  transform:translate(9px, -2px) rotate(2.5deg)
- }
- 4% {
-  transform:translate(8px, 5px) rotate(3.5deg)
- }
- 6% {
-  transform:translate(-3px, 1px) rotate(-1.5deg)
- }
- 8% {
-  transform:translate(0px, -2px) rotate(2.5deg)
- }
- 10% {
-  transform:translate(3px, -4px) rotate(-2.5deg)
- }
- 12% {
-  transform:translate(-1px, 10px) rotate(-.5deg)
- }
- 14% {
-  transform:translate(0px, 10px) rotate(-1.5deg)
- }
- 16% {
-  transform:translate(7px, -1px) rotate(3.5deg)
- }
- 18% {
-  transform:translate(1px, 10px) rotate(3.5deg)
- }
- 20% {
-  transform:translate(1px, -5px) rotate(3.5deg)
- }
- 22% {
-  transform:translate(2px, 1px) rotate(3.5deg)
- }
- 24% {
-  transform:translate(2px, 0px) rotate(2.5deg)
- }
- 26% {
-  transform:translate(2px, 6px) rotate(2.5deg)
- }
- 28% {
-  transform:translate(-7px, 3px) rotate(3.5deg)
- }
- 30% {
-  transform:translate(1px, -1px) rotate(1.5deg)
- }
- 32% {
-  transform:translate(3px, 5px) rotate(-1.5deg)
- }
- 34% {
-  transform:translate(-2px, 1px) rotate(-.5deg)
- }
- 36% {
-  transform:translate(8px, 6px) rotate(.5deg)
- }
- 38% {
-  transform:translate(-2px, 2px) rotate(-2.5deg)
- }
- 40% {
-  transform:translate(7px, -5px) rotate(2.5deg)
- }
- 42% {
-  transform:translate(9px, -9px) rotate(2.5deg)
- }
- 44% {
-  transform:translate(-7px, 0px) rotate(-2.5deg)
- }
- 46% {
-  transform:translate(-5px, -6px) rotate(.5deg)
- }
- 48% {
-  transform:translate(6px, 7px) rotate(.5deg)
- }
- 50% {
-  transform:translate(3px, -9px) rotate(-2.5deg)
- }
- 52% {
-  transform:translate(3px, 9px) rotate(.5deg)
- }
- 54% {
-  transform:translate(-8px, -2px) rotate(-2.5deg)
- }
- 56% {
-  transform:translate(2px, -2px) rotate(-.5deg)
- }
- 58% {
-  transform:translate(-3px, -4px) rotate(3.5deg)
- }
- 60% {
-  transform:translate(0px, 5px) rotate(-2.5deg)
- }
- 62% {
-  transform:translate(-7px, 8px) rotate(.5deg)
- }
- 64% {
-  transform:translate(10px, 6px) rotate(1.5deg)
- }
- 66% {
-  transform:translate(-2px, -6px) rotate(-.5deg)
- }
- 68% {
-  transform:translate(10px, -6px) rotate(2.5deg)
- }
- 70% {
-  transform:translate(0px, 7px) rotate(3.5deg)
- }
- 72% {
-  transform:translate(1px, -8px) rotate(-2.5deg)
- }
- 74% {
-  transform:translate(5px, -1px) rotate(3.5deg)
- }
- 76% {
-  transform:translate(-3px, 9px) rotate(-.5deg)
- }
- 78% {
-  transform:translate(-2px, 8px) rotate(.5deg)
- }
- 80% {
-  transform:translate(7px, 7px) rotate(3.5deg)
- }
- 82% {
-  transform:translate(-2px, -8px) rotate(3.5deg)
- }
- 84% {
-  transform:translate(6px, 5px) rotate(1.5deg)
- }
- 86% {
-  transform:translate(8px, 9px) rotate(2.5deg)
- }
- 88% {
-  transform:translate(10px, 1px) rotate(-2.5deg)
- }
- 90% {
-  transform:translate(9px, -4px) rotate(2.5deg)
- }
- 92% {
-  transform:translate(-8px, -4px) rotate(1.5deg)
- }
- 94% {
-  transform:translate(0px, 8px) rotate(1.5deg)
- }
- 96% {
-  transform:translate(0px, 10px) rotate(1.5deg)
- }
- 98% {
-  transform:translate(-4px, 5px) rotate(2.5deg)
- }
- 0%,100% {
-  transform:translate(0, 0) rotate(0)
- }
+    2% {
+        transform: translate(9px, -2px) rotate(2.5deg)
+    }
+    4% {
+        transform: translate(8px, 5px) rotate(3.5deg)
+    }
+    6% {
+        transform: translate(-3px, 1px) rotate(-1.5deg)
+    }
+    8% {
+        transform: translate(0px, -2px) rotate(2.5deg)
+    }
+    10% {
+        transform: translate(3px, -4px) rotate(-2.5deg)
+    }
+    12% {
+        transform: translate(-1px, 10px) rotate(-.5deg)
+    }
+    14% {
+        transform: translate(0px, 10px) rotate(-1.5deg)
+    }
+    16% {
+        transform: translate(7px, -1px) rotate(3.5deg)
+    }
+    18% {
+        transform: translate(1px, 10px) rotate(3.5deg)
+    }
+    20% {
+        transform: translate(1px, -5px) rotate(3.5deg)
+    }
+    22% {
+        transform: translate(2px, 1px) rotate(3.5deg)
+    }
+    24% {
+        transform: translate(2px, 0px) rotate(2.5deg)
+    }
+    26% {
+        transform: translate(2px, 6px) rotate(2.5deg)
+    }
+    28% {
+        transform: translate(-7px, 3px) rotate(3.5deg)
+    }
+    30% {
+        transform: translate(1px, -1px) rotate(1.5deg)
+    }
+    32% {
+        transform: translate(3px, 5px) rotate(-1.5deg)
+    }
+    34% {
+        transform: translate(-2px, 1px) rotate(-.5deg)
+    }
+    36% {
+        transform: translate(8px, 6px) rotate(.5deg)
+    }
+    38% {
+        transform: translate(-2px, 2px) rotate(-2.5deg)
+    }
+    40% {
+        transform: translate(7px, -5px) rotate(2.5deg)
+    }
+    42% {
+        transform: translate(9px, -9px) rotate(2.5deg)
+    }
+    44% {
+        transform: translate(-7px, 0px) rotate(-2.5deg)
+    }
+    46% {
+        transform: translate(-5px, -6px) rotate(.5deg)
+    }
+    48% {
+        transform: translate(6px, 7px) rotate(.5deg)
+    }
+    50% {
+        transform: translate(3px, -9px) rotate(-2.5deg)
+    }
+    52% {
+        transform: translate(3px, 9px) rotate(.5deg)
+    }
+    54% {
+        transform: translate(-8px, -2px) rotate(-2.5deg)
+    }
+    56% {
+        transform: translate(2px, -2px) rotate(-.5deg)
+    }
+    58% {
+        transform: translate(-3px, -4px) rotate(3.5deg)
+    }
+    60% {
+        transform: translate(0px, 5px) rotate(-2.5deg)
+    }
+    62% {
+        transform: translate(-7px, 8px) rotate(.5deg)
+    }
+    64% {
+        transform: translate(10px, 6px) rotate(1.5deg)
+    }
+    66% {
+        transform: translate(-2px, -6px) rotate(-.5deg)
+    }
+    68% {
+        transform: translate(10px, -6px) rotate(2.5deg)
+    }
+    70% {
+        transform: translate(0px, 7px) rotate(3.5deg)
+    }
+    72% {
+        transform: translate(1px, -8px) rotate(-2.5deg)
+    }
+    74% {
+        transform: translate(5px, -1px) rotate(3.5deg)
+    }
+    76% {
+        transform: translate(-3px, 9px) rotate(-.5deg)
+    }
+    78% {
+        transform: translate(-2px, 8px) rotate(.5deg)
+    }
+    80% {
+        transform: translate(7px, 7px) rotate(3.5deg)
+    }
+    82% {
+        transform: translate(-2px, -8px) rotate(3.5deg)
+    }
+    84% {
+        transform: translate(6px, 5px) rotate(1.5deg)
+    }
+    86% {
+        transform: translate(8px, 9px) rotate(2.5deg)
+    }
+    88% {
+        transform: translate(10px, 1px) rotate(-2.5deg)
+    }
+    90% {
+        transform: translate(9px, -4px) rotate(2.5deg)
+    }
+    92% {
+        transform: translate(-8px, -4px) rotate(1.5deg)
+    }
+    94% {
+        transform: translate(0px, 8px) rotate(1.5deg)
+    }
+    96% {
+        transform: translate(0px, 10px) rotate(1.5deg)
+    }
+    98% {
+        transform: translate(-4px, 5px) rotate(2.5deg)
+    }
+    0%,
+    100% {
+        transform: translate(0, 0) rotate(0)
+    }
 }
 
 /* End csshake */
-