Quick performance optimisation.

From ~240ms generation time to ~95ms.
More optimisation requires changes to generation logic.
Slowest components:
- drawPattern 35ms
	mainly bezier functions
- drawWing 11ms
	mainly mask function
- trimCanvas 28ms