@bleeptrack 2019-06-23
Heya, so sorry for my late reply. I wanted to fix the shirt issue first and now I finally merged your changes, Thanks a lot!